requirements.txt
camelot-py==0.9.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
dnspython==2.3.0
et-xmlfile==1.1.0
idna==3.4
numpy==1.24.2
opencv-python==4.7.0.68
openpyxl==3.1.0
pandas==1.5.3
pdfminer.six==20221105
pycparser==2.21
PyPDF2==2.12.1
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.10.1
six==1.16.0
srvlookup==3.0.0
urllib3==1.26.14
wxcadm>=3.1.0
xmltodict==0.13.0